% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CensusVariableSearch.R
\name{CensusVariableSearch}
\alias{CensusVariableSearch}
\title{CensusVariableSearch}
\usage{
CensusVariableSearch(
  Yr,
  dataset = "acs5",
  variableCode = NULL,
  geog_var = NULL,
  conceptSearch = NULL,
  labelSearch = NULL
)
}
\arguments{
\item{Yr}{The end year of the survey you are pulling data from.}

\item{dataset}{Survey type, which can be "acs5", "acs3", or "acs1". Defailts to "acs5".}

\item{variableCode}{Character string or vector of character strings that can be used to search for a variable if you already know the code. For example, you can check if a code used from a previous survey is still valid.}

\item{geog_var}{Optional string specifying the geography of the variable you are searching for, as this can help narrow results.}

\item{conceptSearch}{Character string or vector of character strings representing a word or phrase found in the "concept" column of the variable table. Useful concept search terms include "Total population", "Median Household Income", "Race".}

\item{labelSearch}{Character string or vector of character strings representing a word or phrase found in the "label" column of the variable table.}
}
\value{
A four column tibble of variables from the requested dataset, filtered by the search terms. If there are no matching search terms a warning will appear and an empty tibble will appear.
}
\description{
This function allows users to search for variables by narrowing down the available options provided in the tidyCensus load_variables function.
}
\details{
The load_variables function allows users to select a dataset and year, but provides no additional means of filtering the thousands of variables that may be present. This function builds on the load_variables function to provide users a way to filter the variable table to reduce the number of variables shown.
}
\examples{
CensusVariableSearch(Yr=2021, dataset="acs5", geog_var = "block group", conceptSearch = "Race")

CensusVariableSearch(Yr=2021, dataset="acs5", variableCode = "B02001_001")

CensusVariableSearch(Yr=2021, dataset="acs5", geog_var = "block group", conceptSearch = "Education")

}
