% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRSEIData.R
\name{GetRSEIData}
\alias{GetRSEIData}
\title{GetRSEIData}
\usage{
GetRSEIData(
  rseiversion = "v2311",
  savetodisk = FALSE,
  directory = NULL,
  seconds = 500
)
}
\arguments{
\item{rseiversion}{Users can currently specify "v2311" (the most recent version of RSEI data) or "v2310" (the last version).}

\item{savetodisk}{Logical. If TRUE, the dataframes are saved as csv files in a folder in the directory provided.}

\item{directory}{An optional path specifying where the data should be saved.}

\item{seconds}{The number of seconds before the download times out. Defaults to 500 but can be extended if necessary.}
}
\value{
A list containing the five RSEI tables. These tables are also assigned to the global environment for use with other functions in this package, and saved as csv files if savetodisk is TRUE.
}
\description{
Function to download and save publicly available data from the EPA's Risk-Screening Environmental Indicators (RSEI) Model.
}
\details{
Five RSEI tables, including the Facility, Chemicals, Submissions, Releases, and Elements are downloaded. See the RSEI Data Dictionary for more information on what is contained in each table.
}
\examples{
\dontrun{RSEIList <- GetRSEIData(rseiversion = "v2311", savetodisk = FALSE)}

}
